requirements.txt

fastapi
uvicorn
pydantic
requests
python-multipart 
sentence-transformers
transformers
torch
faiss-cpu
chromadb
PyPDF2
python-dotenv
groq
pdfplumber
python-dotenv
langchain==0.1.0
langchain-community==0.0.17